backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
cryptography==41.0.7
sqlalchemy==1.4.52
python-dotenv==1.0.0
psycopg2-binary==2.9.9

